Removing compiler warnings

+++ src/hed/libs/data/DataPoint.cpp
@@ -13,7 +13,7 @@ namespace Arc {
   Logger DataPoint::logger(Logger::rootLogger, "DataPoint");
 
   DataPoint::DataPoint(const URL& url) : url(url),
-                                         size(-1),
+                                         size((unsigned long long int)-1),
                                          created(-1),
                                          valid(-1),
                                          triesleft(5) {}
